+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
-# [0.8.2] (2025-09-03)
+# [0.8.3] (2025-09-03)
 
 ## Changes